Refactor gallery feature: rename 'Galary' to 'Gallery', update routes, and implement new photo gallery component with styles
