+++ client/public/data/photos.json
@@ -1,7 +1,9 @@
 [
-  "https://iili.io/3gyOGdx.jpg",
-  "https://iili.io/3gyOWmB.md.jpg",
-  "https://iili.io/3gyOjX1.jpg",
-  "https://iili.io/3gyOhzP.jpg"
-
+  "https://iili.io/3rcaT3g.md.jpg",
+  "https://iili.io/3rccExV.md.jpg",
+  "https://iili.io/3rc0HGI.md.jpg",
+  "https://iili.io/3rc1KKP.md.jpg",
+  "https://iili.io/3rcMp1V.md.jpg",
+  "https://iili.io/3rcXle9.md.jpg",
+  "https://iili.io/3rc8gB1.md.jpg"
 ]
