+++ client/src/pages/Gallery/styles/Gallery.module.scss
@@ -1,5 +1,5 @@
 
-.GalaryTitle {
+.GalleryTitle {
   font-size: 3rem;
   margin-bottom: 1.5rem;
   font-weight: 700;
@@ -19,16 +19,16 @@
 
 
 @media (max-width: 1200px){
-  .GalaryTitle{
+  .GalleryTitle{
     font-size: 3rem;
   }
 }
 @media (max-width: 768px) {
-.GalaryTitle{
+.GalleryTitle{
   font-size: 2.5rem;
 }
 @media (max-width:480px){
-  .GalaryTitle{
+  .GalleryTitle{
     font-size: 2rem;
   }
 }
