+++ client/src/routes/index.jsx
@@ -6,7 +6,7 @@ import Guidelines from './../pages/Guidelines/Guidelines.jsx';
 import Team from './../pages/Team/Team.jsx';
 import FAQ from './../pages/FAQ/FAQ.jsx';
 import Error from './../layouts/Error/Error.jsx';
-import Galary from '../pages/Galary/Galary.jsx';
+import Gallery from '../pages/Gallery/Gallery.jsx';
 
 const router = createBrowserRouter([
     {
@@ -34,8 +34,8 @@ const router = createBrowserRouter([
                 element: <FAQ />
             },
             {
-                path: '/galary',
-                element: <Galary />
+                path: '/gallery',
+                element: <Gallery />
             },
             {
                 path: '*',
